Flesh out reference documentation for the parse_qcodes function
Adds a bit more context to the parse_qcodes function documentation. We need more
complete documentation in order to publish to CRAN.

The updated documentation talks a little bit more about how this function works,
and provides a couple of examples.

I moved the majority of the description into the details. This makes the
description a bit more consise while preserving the details in the appopriately
named details section.

See: #209